[refactor] controller, service 로직 분리
- controller에 try catch 문 추가
- service 로직 분리
