+++ src/views/reservation/equipment/step2.ejs
@@ -120,6 +120,7 @@
           <input type="hidden" name="fromDateTime" value="<%= fromDateTime %>"/>
           <input type="hidden" name="toDateValue" value="<%= toDate %>"/>
           <input type="hidden" name="toDateTime" value="<%= toDateTime %>"/>
+          <input type="hidden" name="equipments" value=" "/>
         </div>
       </form>
     </div>
